requirements.txt
requests >= 2.32.2
fastapi >= 0.111.0
python-escpos >= 3.1